requirements.txt
flask==2.2.2
flask-restful==0.3.9
flask-jwt-extended==4.4.4
Flask-SQLAlchemy==2.5.1
python-dotenv==0.20.0
psycopg2==2.9.3